mux: change maximumIngressQueue to be Int, not Int64
Int is plenty.

+++ network-mux/src/Network/Mux/Types.hs
@@ -37,7 +37,6 @@ import           Prelude hiding (read)
 
 import           Data.Void (Void)
 import           Data.Functor (void)
-import           Data.Int
 import           Data.Ix (Ix (..))
 import           Data.Word
 import qualified Data.ByteString.Lazy as BL
@@ -75,7 +74,7 @@ data MiniProtocolLimits =
        -- | Limit on the maximum number of bytes that can be queued in the
        -- miniprotocol's ingress queue.
        --
-       maximumIngressQueue :: !Int64
+       maximumIngressQueue :: !Int
      }
